Feature/order

+++ src/main/java/com/hanshin/shop/service/CartService.java
@@ -1,6 +1,6 @@
 package com.hanshin.shop.service;
 
-import com.hanshin.shop.entity.cart.CartVO;
+import com.hanshin.shop.vo.cart.CartVO;
 import com.hanshin.shop.repository.CartMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -31,6 +31,10 @@ public class CartService {
         }
     }
 
+    public int count(Long userId) {
+        return cartMapper.count(userId);
+    }
+
     public List<CartVO> findAll(Long userId) {
         return cartMapper.findAll(userId);
     }
@@ -44,7 +48,13 @@ public class CartService {
         return cartMapper.delete(id);
     }
 
-    @Transactional public int deleteAll(Long userId) {
+    @Transactional
+    public int deleteAll(Long userId) {
+        final List<CartVO> all = cartMapper.findAll(userId);
+        log.info("deleteALl {}", all.size());
+        if (all.isEmpty() || all.size() == 0) {
+            throw new IllegalStateException("삭제할 상품이 없습니다.");
+        }
         return cartMapper.deleteAll(userId);
     }
 }
